+++ products/AbstractBuilder.node.js
@@ -65,6 +65,25 @@ class AbstractBuilder extends jtree.TreeNode {
     Disk.write(outputFilePath, transformFn(Disk.read(outputFilePath)))
     this._prettifyFile(outputFilePath)
   }
+  produceProductFromInstructionsTree(productNode, projectRootPath) {
+    const outputFileName = productNode.get("outputFileName")
+    const inputFiles = productNode
+      .getNode("files")
+      .getWordsFrom(1)
+      .map(path => projectRootPath + "/" + path)
+    const firstLine = productNode.get("firstLine") ? productNode.get("firstLine") + "\n" : ""
+    const lastLine = productNode.get("lastLine") ? productNode.get("lastLine") : ""
+    const removeAll = productNode.getNodesByGlobPath("removeAll")
+    const transformFn = code => {
+      removeAll.forEach(node => {
+        code = jtree.Utils.removeAll(code, node.getContent())
+      })
+      return firstLine + code + "\n" + lastLine
+    }
+    if (productNode.getLine() === "browserProduct") this._produceBrowserProductFromTypeScript(inputFiles, outputFileName, transformFn)
+    else this._produceNodeProductFromTypeScript(inputFiles, outputFileName, transformFn)
+    if (productNode.has("executable")) Disk.makeExecutable(projectRootPath + "/products/" + outputFileName)
+  }
   _getProductFolder() {
     return __dirname
   }
@@ -87,6 +106,9 @@ class AbstractBuilder extends jtree.TreeNode {
     packageJson.version = newVersion
     Disk.writeJson(packagePath, packageJson)
     console.log(`Updated ${packagePath} to ${newVersion}`)
+  }
+  buildBuilder() {
+    this._buildTsc(Disk.read(__filename), __dirname + "/builder.js")
   }
   makeGrammarFileTestTree(grammarPath) {
     // todo: can we ditch these dual tests at some point? ideally Grammar should be bootstrapped correct?