+++ products/fire.browser.js
@@ -3,7 +3,7 @@
     createParser() {
       return new jtree.TreeNode.Parser(
         errorNode,
-        Object.assign(Object.assign({}, super.createParser()._getFirstWordMap()), {
+        Object.assign(Object.assign({}, super.createParser()._getFirstWordMapAsObject()), {
           block: blockNode,
           function: functionNode,
           if: ifNode,
@@ -457,7 +457,7 @@ errorNode
     createParser() {
       return new jtree.TreeNode.Parser(
         undefined,
-        Object.assign(Object.assign({}, super.createParser()._getFirstWordMap()), {
+        Object.assign(Object.assign({}, super.createParser()._getFirstWordMapAsObject()), {
           block: blockNode,
           function: functionNode,
           if: ifNode,