+++ products/fire.nodejs.js
@@ -6,7 +6,7 @@
     createParser() {
       return new jtree.TreeNode.Parser(
         errorNode,
-        Object.assign(Object.assign({}, super.createParser()._getFirstWordMap()), {
+        Object.assign(Object.assign({}, super.createParser()._getFirstWordMapAsObject()), {
           block: blockNode,
           function: functionNode,
           if: ifNode,
@@ -460,7 +460,7 @@ errorNode
     createParser() {
       return new jtree.TreeNode.Parser(
         undefined,
-        Object.assign(Object.assign({}, super.createParser()._getFirstWordMap()), {
+        Object.assign(Object.assign({}, super.createParser()._getFirstWordMapAsObject()), {
           block: blockNode,
           function: functionNode,
           if: ifNode,